Add vehicle image in vehicle access page
