--- license-plate-detection/plate-reader-frontend/src/App.css
@@ -1,92 +1,0 @@
-.App {
-  text-align: center;
-  padding: 20px;
-  max-width: 1400px;
-  margin: 0 auto;
-  font-family: Arial, sans-serif;
-}
-
-h1 {
-  color: #2c3e50;
-  margin-bottom: 30px;
-  font-size: 2.5em;
-}
-
-h2 {
-  color: #34495e;
-  margin-bottom: 20px;
-  font-size: 1.8em;
-}
-
-h3 {
-  color: #34495e;
-  margin-bottom: 15px;
-  font-size: 1.4em;
-}
-
-.content {
-  display: flex;
-  gap: 30px;
-  align-items: flex-start;
-}
-
-.video-section {
-  flex: 2;
-}
-
-.detection-section {
-  flex: 1;
-  min-width: 300px;
-}
-
-.video-container {
-  margin-bottom: 20px;
-  position: relative;
-}
-
-.video-container img {
-  max-width: 100%;
-  border-radius: 8px;
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-}
-
-.plate-display {
-  background-color: #f8f9fa;
-  padding: 20px;
-  border-radius: 12px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-  margin-bottom: 20px;
-}
-
-.plate-display h2 {
-  color: #666;
-  margin-bottom: 15px;
-}
-
-.plate-number {
-  font-size: 2em;
-  font-weight: bold;
-  color: #2c3e50;
-  padding: 15px;
-  background-color: white;
-  border-radius: 8px;
-  border: 3px solid #3498db;
-  display: inline-block;
-  min-width: 200px;
-  text-transform: uppercase;
-  letter-spacing: 2px;
-}
-
-.snapshot-container {
-  background-color: #f8f9fa;
-  padding: 20px;
-  border-radius: 12px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-}
-
-.snapshot {
-  max-width: 100%;
-  border-radius: 8px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-  margin-top: 10px;
-}
