--- numberplate-reader/requirements.txt
@@ -1,5 +1,0 @@
-opencv-python
-ultralytics
-easyocr
-numpy
-torch
