+++ plate-detector-service/.gitignore
@@ -1,0 +1,11 @@
+# Ignore snapshots directory
+snapshots/
+
+# Python specific ignores
+__pycache__/
+*.py[cod]
+*$py.class
+*.so
+.Python
+venv/
+ENV/
