+++ plate-detector-service/requirements.txt
@@ -1,0 +1,7 @@
+fastapi==0.68.1
+uvicorn==0.15.0
+opencv-python==4.5.3.56
+numpy==1.21.2
+easyocr==1.4.2
+python-multipart==0.0.5
+requests==2.31.0
