+++ vehicle-detector/main.py
@@ -8,12 +8,14 @@ app = FastAPI()
 class VehicleEntry(BaseModel):
     plate_number: str
     timestamp: datetime = datetime.now()
+    filename: str
 
 @app.post("/detect")
 def detect_vehicle(entry: VehicleEntry):
     data = {
         "plate_number": entry.plate_number,
-        "timestamp": entry.timestamp.isoformat()
+        "timestamp": entry.timestamp.isoformat(),
+        "filename": entry.filename
     }
     publish_vehicle_detected(data)
     return {"message": "Vehicle detected", "data": data}
